+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/DataEntryWrapper/NewEventDataEntryWrapper.component.js
@@ -2,13 +2,16 @@
 
 import * as React from 'react';
 import i18n from '@dhis2/d2-i18n';
+import { Paper } from '@material-ui/core';
 import withStyles from '@material-ui/core/styles/withStyles';
 import { Button } from '@dhis2/ui';
 import DataEntry from './DataEntry/DataEntry.container';
 import EventsList from './RecentlyAddedEventsList/RecentlyAddedEventsList.container';
 import type { ProgramStage, RenderFoundation } from '../../../../metaData';
+import { useScopeTitleText } from '../../../../hooks/useScopeTitleText';
+import { useCurrentTrackedProgramId } from '../../../../hooks/useCurrentProgramId';
 
-const getStyles = () => ({
+const getStyles = ({ typography }) => ({
     flexContainer: {
         display: 'flex',
         alignItems: 'center',
@@ -17,6 +20,17 @@ const getStyles = () => ({
     flexEnd: {
         justifyContent: 'flex-end',
         marginLeft: 'auto',
+    },
+    paper: {
+        marginBottom: typography.pxToRem(10),
+        padding: typography.pxToRem(10),
+    },
+    title: {
+        padding: '8px 0 0px 8px',
+        fontWeight: 500,
+    },
+    marginLeft: {
+        marginLeft: 8,
     },
 });
 
@@ -28,57 +42,61 @@ type Props = {
     stage: ?ProgramStage,
 }
 
+const NewEventDataEntryWrapperPlain = ({
+    classes,
+    formFoundation,
+    formHorizontal,
+    stage,
+    onFormLayoutDirectionChange,
+}: Props) => {
+    let cancelButtonInstance;
 
-class NewEventDataEntryWrapperPlain extends React.Component<Props> {
-    cancelButtonInstance: ?any;
+    const setCancelButtonInstance = (cancelButton: ?any) => {
+        cancelButtonInstance = cancelButton;
+    };
 
-    setCancelButtonInstance = (cancelButtonInstance: ?any) => {
-        this.cancelButtonInstance = cancelButtonInstance;
-    }
+    const programId = useCurrentTrackedProgramId();
+    const titleText = useScopeTitleText(programId);
 
-    render() {
-        const {
-            classes,
-            formFoundation,
-            formHorizontal,
-            stage,
-            onFormLayoutDirectionChange,
-        } = this.props;
+    return (
+        <Paper className={classes.paper}>
+            <div className={classes.title} >
+                New {titleText}
+            </div>
 
-        return (
-            <>
-                <div className={classes.flexContainer}>
-                    <div className={classes.flexEnd}>
-                        {
-                            !formFoundation || formFoundation.customForm ?
-                                null
-                                :
-                                <Button
-                                    onClick={() => onFormLayoutDirectionChange(!formHorizontal)}
-                                    small
-                                >
-                                    {
-                                        formHorizontal
-                                            ?
-                                            i18n.t('Switch to form view')
-                                            :
-                                            i18n.t('Switch to row view')
-                                    }
-                                </Button>
+            <div className={classes.flexContainer}>
+                <div className={classes.flexEnd}>
+                    {
+                        !formFoundation || formFoundation.customForm ?
+                            null
+                            :
+                            <Button
+                                onClick={() => onFormLayoutDirectionChange(!formHorizontal)}
+                                small
+                            >
+                                {
+                                    formHorizontal
+                                        ?
+                                        i18n.t('Switch to form view')
+                                        :
+                                        i18n.t('Switch to row view')
+                                }
+                            </Button>
 
-                        }
-                    </div>
+                    }
                 </div>
+            </div>
+            <div className={classes.marginLeft}>
                 <DataEntry
                     stage={stage}
-                    cancelButtonRef={this.setCancelButtonInstance}
+                    cancelButtonRef={setCancelButtonInstance}
                     formFoundation={formFoundation}
                     formHorizontal={formHorizontal}
                 />
                 <EventsList />
-            </>
-        );
-    }
-}
+            </div>
+        </Paper>
+    );
+};
 
 export const NewEventDataEntryWrapperComponent = withStyles(getStyles)(NewEventDataEntryWrapperPlain);
