+++ src/core_modules/capture-core/components/Pages/New/NewPage.component.js
@@ -4,46 +4,17 @@ import { compose } from 'redux';
 import type { ComponentType } from 'react';
 import withStyles from '@material-ui/core/styles/withStyles';
 import i18n from '@dhis2/d2-i18n';
-import Paper from '@material-ui/core/Paper/Paper';
 import { LockedSelector } from '../../LockedSelector';
 import type { ContainerProps, Props } from './NewPage.types';
 import { withErrorMessageHandler, withLoadingIndicator } from '../../../HOC';
 import { newPageStatuses } from './NewPage.constants';
 import { InefficientSelectionsMessage } from '../../InefficientSelectionsMessage';
-import { TrackedEntityTypeSelector } from '../../TrackedEntityTypeSelector';
-import { scopeTypes } from '../../../metaData';
 import { useScopeInfo } from '../../../hooks/useScopeInfo';
-import { useScopeTitleText } from '../../../hooks/useScopeTitleText';
 import { RegistrationDataEntry } from './RegistrationDataEntry';
 
-const getStyles = ({ typography }) => ({
+const getStyles = () => ({
     container: {
         padding: '24px 24px 16px 24px',
-    },
-    paper: {
-        marginBottom: typography.pxToRem(10),
-        padding: typography.pxToRem(10),
-    },
-    title: {
-        padding: '8px 0 0px 8px',
-        fontWeight: 500,
-    },
-    tetypeContainer: {
-        marginTop: typography.pxToRem(16),
-    },
-    registrationContainer: {
-        marginLeft: typography.pxToRem(8),
-        marginRight: typography.pxToRem(8),
-    },
-    selectorTopMargin: {
-        margin: typography.pxToRem(16),
-    },
-    emptySelectionPaperContent: {
-        display: 'flex',
-        justifyContent: 'center',
-        alignItems: 'center',
-        paddingTop: 50,
-        paddingBottom: 50,
     },
 });
 
@@ -61,7 +32,6 @@ const NewPagePlain = ({
 }: Props) => {
     const { scopeType } = useScopeInfo(currentScopeId);
     const [selectedScopeId, setScopeId] = useState(currentScopeId);
-    const titleText = useScopeTitleText(selectedScopeId);
 
     useEffect(() => {
         setScopeId(currentScopeId);
@@ -84,33 +54,16 @@ const NewPagePlain = ({
         showDefaultViewOnNewPage,
     ]);
 
-    const handleRegistrationScopeSelection = (id) => {
-        setScopeId(id);
-    };
-
-
     return (<>
         <LockedSelector />
         <div data-test="dhis2-capture-registration-page-content" className={classes.container} >
             {
                 newPageStatus === newPageStatuses.DEFAULT &&
-                <Paper className={classes.paper}>
-                    <div className={classes.title} >
-                        New {titleText}
-                    </div>
-                    {
-                        (!scopeType || scopeType === scopeTypes.TRACKED_ENTITY_TYPE) &&
-                        <div className={classes.tetypeContainer}>
-                            <TrackedEntityTypeSelector onSelect={handleRegistrationScopeSelection} />
-                        </div>
-                    }
-                    <div className={classes.registrationContainer}>
-                        <RegistrationDataEntry
-                            dataEntryId={NEW_TEI_DATA_ENTRY_ID}
-                            selectedScopeId={selectedScopeId}
-                        />
-                    </div>
-                </Paper>
+                <RegistrationDataEntry
+                    dataEntryId={NEW_TEI_DATA_ENTRY_ID}
+                    selectedScopeId={selectedScopeId}
+                    setScopeId={setScopeId}
+                />
             }
 
             {
