+++ src/core_modules/capture-core/hooks/useCurrentProgramId.js
@@ -1,0 +1,4 @@
+// @flow
+import { useSelector } from 'react-redux';
+
+export const useCurrentTrackedProgramId = () => useSelector(({ currentSelections }) => currentSelections.programId);
